Language_Translation/requirements.txt
streamlit
google-generativeai
python-dotenv
gTTS
pyperclip